requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
langgraph==0.2.14
google-generativeai==0.8.3
PyPDF2==3.0.1
python-multipart==0.0.9
pytest==8.3.3
python-dotenv==1.0.1